\name{chordplot}
\alias{chordplot}
\title{
Make a chord plot of a result from dendRix
}
\description{
Using the D3 library, generate an interactive chord plot of a dendRix run.
}
\usage{
chordplot(results)
}
\arguments{
  \item{results}{
    Results from a dendRix run (gene/aberration sets); if multiple, use [1]
  }
}
\details{
  Refer to the documentation of the d3.chord package for more specifics.
}
\value{
  The URL of the generated chord plot's webpage.
}
\references{
  None yet.
}
\author{
  Tim Triche, Jr. <tim.triche@gmail.com>
}

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
}
