\name{dendRix-package}
\alias{dendRix-package}
\alias{dendRix}
\docType{package}
\title{
Search for nearly mutually exclusive classes of genetic or other aberrations,
and optionally generate chord plots of the results. 
}
\description{
  see ?dendRix for more 
}
\details{
\tabular{ll}{
Package: \tab dendRix\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-11-26\cr
License: \tab Artistic \cr
}
}
\author{
Maintainer: Tim Triche, Jr. <tim.triche@gmail.com>
}
\references{
  De novo Discovery of Mutated Driver Pathways in Cancer: 
  http://cs.brown.edu/people/braphael/publications/papers/Vandin_etAl_Dendrix_GRpreprint.pdf
}
\keyword{ package }
\examples{
  ## Example to be added ASAP
}
